Rename validators modules and move get_withdrawable_assets
